Move several drinking session utils around
